requirements.txt
astroid==2.2.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-language==1.3.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pylint==2.3.1
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
